+++ go.mod
@@ -1,3 +1,4 @@
 module bravian1/team-shuffler
 
-go 1.23
+//go 1.23
+go 1.22.1
