+++ routes/index.go
@@ -1,0 +1,11 @@
+package routes
+
+import (
+	"text/template"
+	"net/http"
+)
+
+func Index(w http.ResponseWriter, r *http.Request) {
+	tmpl := template.Must(template.ParseFiles("index.html"))
+	tmpl.Execute(w, nil)
+}
