+++ types/entities.go
@@ -1,0 +1,10 @@
+package types
+
+type Player struct {
+	Name string `json:"name"`
+	Role string `json:"role"`
+}
+type Teams struct {
+	Striker  string `json:"striker"`
+	Defender string `json:"defender"`
+}
